rename Archive.py to myarchive.py
